Extract some common color stuff from PopulateSubmissionViewHolder and CreateCardView
